addressing #138 broke out 'Model' class from the specfem tools into its own tool for faster imports SeisFlows command line tool moved heavier imports into functions to speed up call times for other functions that don't need it fixed import statements and all tests pass
