+++ seisflows/optimize/gradient.py
@@ -33,7 +33,7 @@ from seisflows import logger
 from seisflows.tools import msg, unix
 from seisflows.tools.config import Dict
 from seisflows.tools.math import angle, dot
-from seisflows.tools.specfem import Model
+from seisflows.tools.model import Model
 from seisflows.plugins import line_search as line_search_dir
