+++ seisflows/workflow/migration.py
@@ -21,7 +21,7 @@ import os
 
 from seisflows import logger
 from seisflows.tools import msg, unix
-from seisflows.tools.specfem import Model
+from seisflows.tools.model import Model
 from seisflows.workflow.forward import Forward
